refactor FSC to not use fft utils and add tests
