+++ src/libtilt/shapes/geometry_utils.py
@@ -1,3 +1,4 @@
+import einops
 import torch
 
 
@@ -5,9 +6,10 @@ def _angle_between_vectors(a: torch.Tensor, b: torch.Tensor) -> torch.Tensor:
     """a can optionally be an nD stack, return in degrees"""
     a = torch.as_tensor(a)
     b = torch.as_tensor(b, device=a.device, dtype=a.dtype)
-    n_stack_dims = a.ndim - 1
-    stack_dims = _unique_characters(n_stack_dims)
-    angles = torch.arccos(torch.einsum(f'{stack_dims}v,v->{stack_dims}', a, b))
+    a, ps = einops.pack([a], pattern='* vec')
+    dot_products = einops.einsum(a, b, 'b vec, vec -> b')
+    angles = torch.rad2deg(torch.arccos(dot_products))
+    [angles] = einops.unpack(angles, packed_shapes=ps, pattern='*')
     return torch.rad2deg(angles)
